+++ spec/controllers/cards_controller_spec.rb
@@ -1,10 +1,6 @@
 require 'rails_helper'
 
 describe CardsController do
-  subject(:do_request) do
-    get :show, format: 'json', owner_type: object.class.name.underscore, owner_id: object.id
-  end
-
   let(:user) { FactoryGirl.create(:user) }
   let(:my_journal) { FactoryGirl.create(:journal) }
   let(:my_other_journal) { FactoryGirl.create(:journal) }
@@ -57,37 +53,81 @@ describe CardsController do
   end
 
   describe "#show" do
-    let(:user) { create :user, :site_admin }
-
-    before do
-      stub_sign_in user
+    subject(:do_request) do
+      get :show, format: 'json', id: card.id
     end
+    let(:card) { FactoryGirl.create(:card) }
 
-    context "resource is answerable" do
-      let(:card) { FactoryGirl.create(:card) }
-      let(:object) { FactoryGirl.create(:cover_letter_task, card: card) }
+    it_behaves_like 'an unauthenticated json request'
 
-      it "returns a serialized card" do
-        do_request
-        expect(response.status).to eq(200)
-        expect(res_body).to include("card")
+    context 'and the user is signed in' do
+      context "when the user does not have access" do
+        before do
+          stub_sign_in(user)
+          allow(user).to receive(:can?)
+            .with(:view, card)
+            .and_return false
+          do_request
+        end
+
+        it { is_expected.to responds_with(403) }
       end
-    end
 
-    context "resource is not answerable" do
-      let(:object) { FactoryGirl.create(:user) }
+      context 'user has access' do
+        before do
+          stub_sign_in user
+          allow(user).to receive(:can?).with(:view, card).and_return(true)
+        end
 
-      it "returns a 422" do
-        do_request
-        expect(response.status).to eq(422)
+        it { is_expected.to responds_with 200 }
+
+        it 'returns the serialized card' do
+          do_request
+          expect(res_body['card']['id']).to be card.id
+        end
       end
     end
   end
 
-  context "authentication" do
-    let(:card) { FactoryGirl.create(:card) }
-    let(:object) { FactoryGirl.create(:cover_letter_task, card: card) }
+  describe "#create" do
+    subject(:do_request) do
+      post(:create, format: 'json', card: {
+             name: name,
+             journal_id: my_journal.id
+           })
+    end
+    let(:name) { "Steve" }
 
     it_behaves_like 'an unauthenticated json request'
+
+    context 'and the user is signed in' do
+      context "when the user does not have access" do
+        before do
+          stub_sign_in(user)
+          allow(user).to receive(:can?)
+            .with(:create_card, my_journal)
+            .and_return false
+          do_request
+        end
+
+        it { is_expected.to responds_with(403) }
+      end
+
+      context 'user has access' do
+        before do
+          stub_sign_in user
+          allow(user).to receive(:can?)
+            .with(:create_card, my_journal)
+            .and_return(true)
+        end
+
+        it { is_expected.to responds_with 201 }
+
+        it 'returns the serialized card' do
+          do_request
+          expect(res_body['card']['name']).to eq name
+        end
+      end
+    end
   end
 end