+++ app/services/paper_converters/identity_paper_converter.rb
@@ -1,0 +1,8 @@
+module PaperConverters
+  # A paper 'converter' that provides the url of the original document
+  class IdentityPaperConverter < RedirectingPaperConverter
+    def download_url
+      Attachment.authenticated_url_for_key(@versioned_text.s3_full_path)
+    end
+  end
+end