+++ spec/models/versioned_text_spec.rb
@@ -76,12 +76,12 @@ describe VersionedText do
     it "has matching file and versioned_text s3 directories" do
         paper.draft.be_minor_version!
         expect(paper.file.s3_dir).not_to be_nil
-        expect(paper.file.s3_dir).to eq(versioned_text.s3_dir)
+        expect(paper.file.s3_dir).to eq(versioned_text.manuscript_s3_path)
       end
 
       it "has matching file and versioned_text filenames" do
         paper.draft.be_minor_version!
-        expect(paper.file[:file]).to eq(versioned_text.file)
+        expect(paper.file[:file]).to eq(versioned_text.manuscript_filename)
       end
   end
 
@@ -110,12 +110,12 @@ describe VersionedText do
     it "has matching file and versioned_text s3 directories" do
         paper.draft.be_major_version!
         expect(paper.file.s3_dir).not_to be_nil
-        expect(paper.file.s3_dir).to eq(versioned_text.s3_dir)
+        expect(paper.file.s3_dir).to eq(versioned_text.manuscript_s3_path)
       end
 
       it "has matching file and versioned_text filenames" do
         paper.draft.be_major_version!
-        expect(paper.file[:file]).to eq(versioned_text.file)
+        expect(paper.file[:file]).to eq(versioned_text.manuscript_filename)
       end
   end
 
@@ -131,11 +131,11 @@ describe VersionedText do
 
       it "has matching file and versioned_text s3 directories" do
         expect(paper.file.s3_dir).not_to be_nil
-        expect(paper.file.s3_dir).to eq(versioned_text.s3_dir)
+        expect(paper.file.s3_dir).to eq(versioned_text.manuscript_s3_path)
       end
 
       it "has matching file and versioned_text filenames" do
-        expect(paper.file[:file]).to eq(versioned_text.file)
+        expect(paper.file[:file]).to eq(versioned_text.manuscript_filename)
       end
 
       it "fails" do