+++ CHANGELOG.md
@@ -9,6 +9,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 * The manuscript upload card prompts for a backing source file if a PDF manuscript is uploaded.
 * Source files show can be downloaded
 * Added "Confirm Authorship" button to email to co authors.
+* Added PDF submissions for PLOS Biology
 ### Changed
 * Changed new manuscript filetype text for pdf-enabled journals to mention the need for a sourcefile upload backing pdf manuscripts
 * Temporarily disabled coauthor notification